Remove optimization flags. Never for main branch
This commit is to be used when OCaml is compiled with optimization flags (ex. DkSDK CMake).

Optimization flags like -O2 will cause MSVC failures if OCaml was compiled with the -RTC1 flag (cl : Command line error D8016 : '/RTC1' and '/O2' command-line options are incompatible). And even if it were not incompatible, the optimization flags that OCaml was compiled with should always be used.